+++ fdmax_windows.go
@@ -4,10 +4,10 @@ package fdmax
 
 // Get the current limits
 func Get() (*Limits, error) {
-	return nil, UnsupportedPlatformError
+	return nil, ErrUnsupportedPlatform
 }
 
 // Set new system limits
 func Set(maxLimit uint64) error {
-	return UnsupportedPlatformError
+	return ErrUnsupportedPlatform
 }
